+++ test/ModularPipelines.UnitTests/Helpers/FtpTests.cs
@@ -29,11 +29,11 @@ public class FtpTests : TestBase
 
         var fileContents = await localPath.ReadAsync();
         
-        await Assert.Multiple(() =>
+        await using (Assert.Multiple())
         {
-            Assert.That(response).Is.EqualTo(FtpStatus.Success);
-            Assert.That(fileContents).Does.StartWith("6jack");
-        });
+            await Assert.That(response).Is.EqualTo(FtpStatus.Success);
+            await Assert.That(fileContents).Does.StartWith("6jack");
+        }
     }
 
     [Test]