+++ test/ModularPipelines.UnitTests/Helpers/InstallerTests.cs
@@ -3,7 +3,7 @@ using ModularPipelines.Options;
 using ModularPipelines.Options.Linux;
 using ModularPipelines.Options.Windows;
 using ModularPipelines.TestHelpers;
-using TUnit.Assertions.Extensions.Is;
+using TUnit.Assertions.Extensions.Numbers;
 using TUnit.Core.Exceptions;
 
 namespace ModularPipelines.UnitTests.Helpers;