+++ test/ModularPipelines.UnitTests/KeyValueTests.cs
@@ -10,11 +10,11 @@ public class KeyValueTests
     {
         KeyValue keyValue = ("one", "two");
         
-        await Assert.Multiple(() =>
+        await using (Assert.Multiple())
         {
-            Assert.That(keyValue.Key).Is.EqualTo("one");
-            Assert.That(keyValue.Value).Is.EqualTo("two");
-        });
+            await Assert.That(keyValue.Key).Is.EqualTo("one");
+            await Assert.That(keyValue.Value).Is.EqualTo("two");
+        }
     }
 
     [Test]
@@ -22,11 +22,11 @@ public class KeyValueTests
     {
         KeyValue keyValue = new Tuple<string, string>("one", "two");
         
-        await Assert.Multiple(() =>
+        await using (Assert.Multiple())
         {
-            Assert.That(keyValue.Key).Is.EqualTo("one");
-            Assert.That(keyValue.Value).Is.EqualTo("two");
-        });
+            await Assert.That(keyValue.Key).Is.EqualTo("one");
+            await Assert.That(keyValue.Value).Is.EqualTo("two");
+        }
     }
 
     [Test]
@@ -34,10 +34,10 @@ public class KeyValueTests
     {
         KeyValue keyValue = new KeyValuePair<string, string>("one", "two");
         
-        await Assert.Multiple(() =>
+        await using (Assert.Multiple())
         {
-            Assert.That(keyValue.Key).Is.EqualTo("one");
-            Assert.That(keyValue.Value).Is.EqualTo("two");
-        });
+            await Assert.That(keyValue.Key).Is.EqualTo("one");
+            await Assert.That(keyValue.Value).Is.EqualTo("two");
+        }
     }
 }