+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,2 +1,1 @@
-*   `[DependsOnAllModulesInheritingFromAttribute]` to allow a module to wait on all modules inheriting from a base without needing to list each module explicitly
-*   Better JSON serialization of `Folder` objects
+*   .NET 7 build target has been removed as it is end of life